+++ src/quotes/QuoteHome.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import { Table } from 'reactstrap';
+import QuoteUpdater from './UpdateQuotes';
 
 type QHProps = {
     userToken: string
 }
 
 type QHState = {
-    myQuotes: Quote[]
+    myQuotes: Quote[],
+    updateActive: boolean,
+    quoteToUpdate: {}
 }
 
 type Quote = {
@@ -18,7 +21,9 @@ export default class QuoteHome extends React.Component<QHProps, QHState> {
     constructor(props: QHProps){
         super(props)
         this.state = {
-            myQuotes: []
+            myQuotes: [],
+            updateActive: false,
+            quoteToUpdate: {}
         }
     }
 
@@ -44,21 +49,36 @@ export default class QuoteHome extends React.Component<QHProps, QHState> {
     }
 
     deleteQuote = async (quote: any) => {
-        const deleteURL = `http://localhost:3000/quote/${quote.id}`
-        try {
-            const byeQuote = await fetch (deleteURL, {
-                method: "DELETE",
-                headers: new Headers({
-                    'Content-Type': 'application/json',
-                    'Authorization': `Bearer ${this.props.userToken}`
+        const confirm = prompt("Are you sure you want to delete this quote?", "Yes")
+        if (confirm){
+            const deleteURL = `http://localhost:3000/quote/${quote.id}`
+            try {
+                const byeQuote = await fetch (deleteURL, {
+                    method: "DELETE",
+                    headers: new Headers({
+                        'Content-Type': 'application/json',
+                        'Authorization': `Bearer ${this.props.userToken}`
+                    })
                 })
-            })
-            console.log(byeQuote)
-            this.viewMyQuotes()
-        } catch (err) {
-            console.log(err)
+                console.log(byeQuote)
+                this.viewMyQuotes()
+            } catch (err) {
+                console.log(err)
+            }
         }
     }
+
+    editUpdateQuote = (quote: {}) => {
+        this.setState({quoteToUpdate: quote})
+    };
+
+    updateOn = () => {
+        this.setState({updateActive: true})
+    };
+
+    updateOff = () => {
+        this.setState({updateActive: false})
+    };
 
     quoteMapper = (): JSX.Element[] => {
         return this.state.myQuotes.map((quote: Quote, index: number) => {
@@ -66,7 +86,7 @@ export default class QuoteHome extends React.Component<QHProps, QHState> {
                     <tbody>
                         <tr key={index}>
                             <td>{quote.quoteBody}</td>
-                            <td><button value={quote.id}>Update</button></td>
+                            <td><button value={quote.id} onClick={() => {this.editUpdateQuote(quote.id)}} >Update</button></td>
                             <td><button onClick={() => {this.deleteQuote(quote)}}>Delete</button></td>
                         </tr>
                     </tbody>
@@ -78,13 +98,11 @@ export default class QuoteHome extends React.Component<QHProps, QHState> {
         return(
             <div>
                 <div>
-                    <h1><button style={{fontSize: "xx-large"}}>&#x1F49F; Add a Quote &#x1F49F;</button></h1>
-                </div>
-                <div>
                 <h1>Quotes I've added:</h1>
                 <Table>
                 {this.quoteMapper()}
                 </Table>
+                {this.state.updateActive ? <QuoteUpdater toBeUpdated={this.state.quoteToUpdate} updateOff={this.updateOff} token={this.props.userToken} viewMyQuotes={this.viewMyQuotes} /> : <></>}
                 </div>
             </div>
         )
